finish

+++ mandatory/events.c
@@ -6,7 +6,7 @@
 /*   By: aychikhi <aychikhi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/10 15:48:09 by aychikhi          #+#    #+#             */
-/*   Updated: 2025/01/15 11:21:39 by aychikhi         ###   ########.fr       */
+/*   Updated: 2025/01/16 11:57:27 by aychikhi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,8 +50,8 @@ int	mouse_fun(int button, int x, int y, t_fractal *fractol)
 	double			mouse_y;
 	const double	zoom_factor = 0.1;
 
-	mouse_x = fractol->shift_x + (x * (fractol->zoom / WIDTH));
-	mouse_y = fractol->shift_y + (y * (fractol->zoom / HEIGHT));
+	mouse_x = fractol->shift_x + ((double)x * (fractol->zoom / WIDTH));
+	mouse_y = fractol->shift_y + ((double)y * (fractol->zoom / HEIGHT));
 	if (button == 4)
 	{
 		fractol->zoom *= (1 - zoom_factor);
